Fix fire/ice logic race condition bug
Symptom: if a Ball is on fire or frozen, it stays this way even on collision with the opposing type.

Problem: race condition - eg. if it's on fire, it unfreezes the block so the block doesn't extinguish it

Solution: OnCollisionEnter, queue up the relevant state change command, and then invoke it OnCollisionExit.