(WIP)
Try to make QuestionBlock randomly change the type of all blocks currently in play, when hit

There is a bug where sometimes Loads of blocks are created in an infinite loop and Unity crashes

I have spent too long on this so I will give up.

Also in this commit are various tidying and debugging efforts made towards this cause.